requirements.txt
Flask==3.0.2
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.1
Pillow==10.2.0 